oepnhands-feature-intelligent-agent-evaluation-system/oepnhands-feature-intelligent-agent-evaluation-system/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
aiofiles
python-docx
openpyxl
pandas
requests
scikit-learn
jieba
numpy
aiosqlite
aiohttp